Rename wireframe.py as wireframe_toroidal.py for specificity
Also update references in related files for consistency
