+++ src/simsopt/geo/__init__.py
@@ -23,7 +23,7 @@ from .surfacerzfourier import *
 from .surfacexyzfourier import *
 from .surfacexyztensorfourier import *
 from .strain_optimization import *
-from .wireframe import *
+from .wireframe_toroidal import *
 from .ports import *
 
 from .permanent_magnet_grid import *
@@ -39,4 +39,5 @@ __all__ = (curve.__all__ + curvehelical.__all__ +
            surfacerzfourier.__all__ + surfacexyzfourier.__all__ +
            surfacexyztensorfourier.__all__ + surfaceobjectives.__all__ +
            strain_optimization.__all__ + framedcurve.__all__ + 
-           wireframe.__all__ + ports.__all__ + permanent_magnet_grid.__all__)
+           wireframe_toroidal.__all__ + ports.__all__ + 
+           permanent_magnet_grid.__all__)
