+++ tests/geo/test_wireframe_toroidal.py
